Move vscode.newWindow & vscode.removeFromRecentlyOpened out of apiCommands. For #110583
