+++ src/vs/workbench/api/common/apiCommands.ts
@@ -7,8 +7,6 @@ import { URI } from 'vs/base/common/uri';
 import * as typeConverters from 'vs/workbench/api/common/extHostTypeConverters';
 import { CommandsRegistry, ICommandService, ICommandHandler } from 'vs/platform/commands/common/commands';
 import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
-import { IOpenEmptyWindowOptions } from 'vs/platform/windows/common/windows';
-import { IWorkspacesService, IRecent } from 'vs/platform/workspaces/common/workspaces';
 import { ILogService } from 'vs/platform/log/common/log';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { IViewDescriptorService, IViewsService, ViewVisibilityState } from 'vs/workbench/common/views';
@@ -29,83 +27,6 @@ function adjustHandler(handler: (executor: ICommandsExecutor, ...args: any[]) =>
 		return handler(accessor.get(ICommandService), ...args);
 	};
 }
-
-interface INewWindowAPICommandOptions {
-	reuseWindow?: boolean;
-	/**
-	 * If set, defines the remoteAuthority of the new window. `null` will open a local window.
-	 * If not set, defaults to remoteAuthority of the current window.
-	 */
-	remoteAuthority?: string | null;
-}
-
-export class NewWindowAPICommand {
-	public static readonly ID = 'vscode.newWindow';
-	public static execute(executor: ICommandsExecutor, options?: INewWindowAPICommandOptions): Promise<any> {
-		const commandOptions: IOpenEmptyWindowOptions = {
-			forceReuseWindow: options && options.reuseWindow,
-			remoteAuthority: options && options.remoteAuthority
-		};
-
-		return executor.executeCommand('_files.newWindow', commandOptions);
-	}
-}
-CommandsRegistry.registerCommand({
-	id: NewWindowAPICommand.ID,
-	handler: adjustHandler(NewWindowAPICommand.execute),
-	description: {
-		description: 'Opens an new window',
-		args: [
-		]
-	}
-});
-
-CommandsRegistry.registerCommand('_workbench.removeFromRecentlyOpened', function (accessor: ServicesAccessor, uri: URI) {
-	const workspacesService = accessor.get(IWorkspacesService);
-	return workspacesService.removeRecentlyOpened([uri]);
-});
-
-export class RemoveFromRecentlyOpenedAPICommand {
-	public static readonly ID = 'vscode.removeFromRecentlyOpened';
-	public static execute(executor: ICommandsExecutor, path: string | URI): Promise<any> {
-		if (typeof path === 'string') {
-			path = path.match(/^[^:/?#]+:\/\//) ? URI.parse(path) : URI.file(path);
-		} else {
-			path = URI.revive(path); // called from extension host
-		}
-		return executor.executeCommand('_workbench.removeFromRecentlyOpened', path);
-	}
-}
-CommandsRegistry.registerCommand(RemoveFromRecentlyOpenedAPICommand.ID, adjustHandler(RemoveFromRecentlyOpenedAPICommand.execute));
-
-interface RecentEntry {
-	uri: URI;
-	type: 'workspace' | 'folder' | 'file';
-	label?: string;
-	remoteAuthority?: string;
-}
-
-CommandsRegistry.registerCommand('_workbench.addToRecentlyOpened', async function (accessor: ServicesAccessor, recentEntry: RecentEntry) {
-	const workspacesService = accessor.get(IWorkspacesService);
-	let recent: IRecent | undefined = undefined;
-	const uri = recentEntry.uri;
-	const label = recentEntry.label;
-	const remoteAuthority = recentEntry.remoteAuthority;
-	if (recentEntry.type === 'workspace') {
-		const workspace = await workspacesService.getWorkspaceIdentifier(uri);
-		recent = { workspace, label, remoteAuthority };
-	} else if (recentEntry.type === 'folder') {
-		recent = { folderUri: uri, label, remoteAuthority };
-	} else {
-		recent = { fileUri: uri, label, remoteAuthority };
-	}
-	return workspacesService.addRecentlyOpened([recent]);
-});
-
-CommandsRegistry.registerCommand('_workbench.getRecentlyOpened', async function (accessor: ServicesAccessor) {
-	const workspacesService = accessor.get(IWorkspacesService);
-	return workspacesService.getRecentlyOpened();
-});
 
 CommandsRegistry.registerCommand('_extensionTests.setLogLevel', function (accessor: ServicesAccessor, level: number) {
 	const logService = accessor.get(ILogService);
