+++ src/vs/workbench/browser/actions/workspaceCommands.ts
@@ -20,8 +20,8 @@ import { IModeService } from 'vs/editor/common/services/modeService';
 import { IFileDialogService, IPickAndOpenOptions } from 'vs/platform/dialogs/common/dialogs';
 import { URI } from 'vs/base/common/uri';
 import { Schemas } from 'vs/base/common/network';
-import { IOpenWindowOptions, IWindowOpenable } from 'vs/platform/windows/common/windows';
-import { hasWorkspaceFileExtension } from 'vs/platform/workspaces/common/workspaces';
+import { IOpenEmptyWindowOptions, IOpenWindowOptions, IWindowOpenable } from 'vs/platform/windows/common/windows';
+import { hasWorkspaceFileExtension, IRecent, IWorkspacesService } from 'vs/platform/workspaces/common/workspaces';
 import { IPathService } from 'vs/workbench/services/path/common/pathService';
 import { ILocalizedString } from 'vs/platform/actions/common/actions';
 
@@ -184,4 +184,92 @@ CommandsRegistry.registerCommand({
 			}
 		]
 	}
+});
+
+interface INewWindowAPICommandOptions {
+	reuseWindow?: boolean;
+	/**
+	 * If set, defines the remoteAuthority of the new window. `null` will open a local window.
+	 * If not set, defaults to remoteAuthority of the current window.
+	 */
+	remoteAuthority?: string | null;
+}
+
+CommandsRegistry.registerCommand({
+	id: 'vscode.newWindow',
+	handler: (accessor: ServicesAccessor, options?: INewWindowAPICommandOptions) => {
+		const commandOptions: IOpenEmptyWindowOptions = {
+			forceReuseWindow: options && options.reuseWindow,
+			remoteAuthority: options && options.remoteAuthority
+		};
+		const commandService = accessor.get(ICommandService);
+		return commandService.executeCommand('_files.newWindow', commandOptions);
+	},
+	description: {
+		description: 'Opens an new window depending on the newWindow argument.',
+		args: [
+			{
+				name: 'options',
+				description: '(optional) Options. Object with the following properties: ' +
+					'`reuseWindow`: Whether to open a new window or the same. Defaults to opening in a new window. ',
+				constraint: (value: any) => value === undefined || typeof value === 'object'
+			}
+		]
+	}
+});
+
+// recent history commands
+
+CommandsRegistry.registerCommand('_workbench.removeFromRecentlyOpened', function (accessor: ServicesAccessor, uri: URI) {
+	const workspacesService = accessor.get(IWorkspacesService);
+	return workspacesService.removeRecentlyOpened([uri]);
+});
+
+
+CommandsRegistry.registerCommand({
+	id: 'vscode.removeFromRecentlyOpened',
+	handler: (accessor: ServicesAccessor, path: string | URI): Promise<any> => {
+		if (typeof path === 'string') {
+			path = path.match(/^[^:/?#]+:\/\//) ? URI.parse(path) : URI.file(path);
+		} else {
+			path = URI.revive(path); // called from extension host
+		}
+		const workspacesService = accessor.get(IWorkspacesService);
+		return workspacesService.removeRecentlyOpened([path]);
+	},
+	description: {
+		description: 'Removes an entry with the given path from the recently opened list.',
+		args: [
+			{ name: 'path', description: 'URI or URI string to remove from recently opened.', constraint: (value: any) => typeof value === 'string' || value instanceof URI }
+		]
+	}
+});
+
+interface RecentEntry {
+	uri: URI;
+	type: 'workspace' | 'folder' | 'file';
+	label?: string;
+	remoteAuthority?: string;
+}
+
+CommandsRegistry.registerCommand('_workbench.addToRecentlyOpened', async function (accessor: ServicesAccessor, recentEntry: RecentEntry) {
+	const workspacesService = accessor.get(IWorkspacesService);
+	let recent: IRecent | undefined = undefined;
+	const uri = recentEntry.uri;
+	const label = recentEntry.label;
+	const remoteAuthority = recentEntry.remoteAuthority;
+	if (recentEntry.type === 'workspace') {
+		const workspace = await workspacesService.getWorkspaceIdentifier(uri);
+		recent = { workspace, label, remoteAuthority };
+	} else if (recentEntry.type === 'folder') {
+		recent = { folderUri: uri, label, remoteAuthority };
+	} else {
+		recent = { fileUri: uri, label, remoteAuthority };
+	}
+	return workspacesService.addRecentlyOpened([recent]);
+});
+
+CommandsRegistry.registerCommand('_workbench.getRecentlyOpened', async function (accessor: ServicesAccessor) {
+	const workspacesService = accessor.get(IWorkspacesService);
+	return workspacesService.getRecentlyOpened();
 });
